@@ -1,23 +1,34 @@
-const lightCodeTheme = require('prism-react-renderer/themes/github');
-const darkCodeTheme = require('prism-react-renderer/themes/dracula');
+// @ts-check
+// `@type` JSDoc annotations allow editor autocompletion and type checking
+// (when paired with `@ts-check`).
+// There are various equivalent ways to declare your Docusaurus config.
+// See: https://docusaurus.io/docs/api/docusaurus-config
 
-/** @type {import('@docusaurus/types').DocusaurusConfig} */
-module.exports = {
+import {themes as prismThemes} from 'prism-react-renderer';
+
+/** @type {import('@docusaurus/types').Config} */
+const config = {
   title: 'MadMachine',
   tagline: 'MadMachine Documentation',
   favicon: 'img/favicon.ico',
 
+  // Set the production url of your site here
   url: 'https://docs.madmachine.io',
+  // Set the /<baseUrl>/ pathname under which your site is served
+  // For GitHub pages deployment, it is often '/<projectName>/'
   baseUrl: '/',
 
+  // GitHub pages deployment config.
+  // If you aren't using GitHub pages, you don't need these.
   organizationName: 'MadMachine', // Usually your GitHub org/user name.
   projectName: 'docs', // Usually your repo name.
 
   onBrokenLinks: 'throw',
   onBrokenMarkdownLinks: 'warn',
-  // Even if you don't use internalization, you can use this field to set useful
-  // metadata like html lang. For example, if your site is Chinese, you may want
-  // to replace "en" with "zh-Hans".
+
+  // Even if you don't use internationalization, you can use this field to set
+  // useful metadata like html lang. For example, if your site is Chinese, you
+  // may want to replace "en" with "zh-Hans".
   i18n: {
     defaultLocale: 'en',
     locales: ['en'],
@@ -26,21 +37,33 @@ module.exports = {
   presets: [
     [
       'classic',
+      /** @type {import('@docusaurus/preset-classic').Options} */
       ({
         docs: {
-          sidebarPath: require.resolve('./sidebars.js'),
+          sidebarPath: './sidebars.js',
           // Please change this to your repo.
+          // Remove this to remove the "edit this page" links.
           editUrl:
             'https://github.com/madmachineio/docs/',
           routeBasePath: '/',
         },
         blog: {
-          blogTitle: 'Docusaurus blog!',
-          blogDescription: 'A Docusaurus powered blog!',
-          postsPerPage: 'ALL',
+          showReadingTime: true,
+          feedOptions: {
+            type: ['rss', 'atom'],
+            xslt: true,
+          },
+          // Please change this to your repo.
+          // Remove this to remove the "edit this page" links.
+          editUrl:
+            'https://github.com/madmachineio/docs/',
+          // Useful options to enforce blogging best practices
+          onInlineTags: 'warn',
+          onInlineAuthors: 'warn',
+          onUntruncatedBlogPosts: 'warn',
         },
         theme: {
-          customCss: require.resolve('./src/css/custom.css'),
+          customCss: './src/css/custom.css',
         },
         gtag: {
           // You can also use your "G-" Measurement ID here.
@@ -53,55 +76,57 @@ module.exports = {
   ],
 
   themeConfig:
-  /** @type {import('@docusaurus/preset-classic').ThemeConfig} */
-  ({
-    navbar: {
-      //title: 'Machine',
-      style: 'dark',
-      hideOnScroll: false,
-      logo: {
-        alt: 'MadMachine',
-        src: 'img/logo_dark.svg',
-        srcDark: 'img/logo_dark.svg',
-        href: 'https://madmachine.io'
-      },
-      items: [
-        {
-          type: 'doc',
-          docId: 'overview/introduction',
-          position: 'left',
-          label: 'Overview',
+    /** @type {import('@docusaurus/preset-classic').ThemeConfig} */
+    ({
+      // Replace with your project's social card
+      // image: 'img/docusaurus-social-card.jpg',
+      navbar: {
+        // title: 'MadMachine',
+        style: 'dark',
+        hideOnScroll: false,
+        logo: {
+          alt: 'MadMachine',
+          src: 'img/logo_dark.svg',
+          srcDark: 'img/logo_dark.svg',
+          href: 'https://madmachine.io'
         },
-        {
-          type: 'doc',
-          docId: 'learn/introduction',
-          position: 'left',
-          label: 'Learning Playground',
-        },
-        // {
-        //   type: 'doc',
-        //   docId: 'projects/overview',
-        //   position: 'left',
-        //   label: 'Projects',
-        // },
-        {
-          type: 'doc',
-          docId: 'hardware/swiftio-micro',
-          position: 'left',
-          label: 'Hardware',
-        },
-        {
-          type: 'doc',
-          docId: 'reference/reference',
-          position: 'left',
-          label: 'API Reference',
-        },
-        {to: '/blog', label: 'Blog', position: 'left'},
-        {
-          href: 'https://github.com/madmachineio/docs',
-          label: 'GitHub',
-          position: 'right',
-        },
+        items: [
+          {
+            type: 'doc',
+            docId: 'overview/introduction',
+            position: 'left',
+            label: 'Overview',
+          },
+          {
+            type: 'doc',
+            docId: 'learn/introduction',
+            position: 'left',
+            label: 'Learning Playground',
+          },
+          // {
+          //   type: 'doc',
+          //   docId: 'projects/overview',
+          //   position: 'left',
+          //   label: 'Projects',
+          // },
+          {
+            type: 'doc',
+            docId: 'hardware/swiftio-micro',
+            position: 'left',
+            label: 'Hardware',
+          },
+          {
+            type: 'doc',
+            docId: 'reference/reference',
+            position: 'left',
+            label: 'API Reference',
+          },
+          {to: '/blog', label: 'Blog', position: 'left'},
+          {
+            href: 'https://github.com/madmachineio/docs',
+            label: 'GitHub',
+            position: 'right',
+          },
       ],
     },
     docs: {
@@ -143,11 +168,11 @@ module.exports = {
         },
       ],
     },
-    prism: {
-      theme: require('prism-react-renderer/themes/github'),
-      darkTheme: require("prism-react-renderer/themes/vsDark"),
-      additionalLanguages: ['swift'],
-    },
+      prism: {
+        theme: prismThemes.github,
+        darkTheme: prismThemes.dracula,
+        additionalLanguages: ['swift'],
+      },
     algolia: {
       // If Algolia did not provide you any appId, use 'BH4D9OD16A'
       appId: '6R1F0BVDSY',
@@ -166,9 +191,11 @@ module.exports = {
       // Optional: Algolia search parameters
       // searchParameters: {},
     },
-  }),
+    }),
   stylesheets: [
     'https://fonts.googleapis.com/css?family=Roboto:400,400i,700',
     'https://fonts.googleapis.com/css?family=Source+Code+Pro',
   ],
 };
+
+export default config;
